Separate tests into files per resource
